Split out validation helpers
